Refactor: 시청 도메인 컨트롤러 리팩토링
- 페이지를 불러오는 메서드를 별도의 클래스로 분리
